refactor: update TextSection exports and remove unused components
