+++ src/features/UI/index.js
@@ -1,10 +1,8 @@
 export { default as Tag } from "./Tag/Tag";
 export { default as Section } from "./Section/Section";
-export {
-  default as TextSection,
-  TextSectionBraceAside,
-  TextSectionQuoteAside,
-} from "./TextSection/TextSection";
+export { default as TextSection } from "./TextSection/TextSection";
+export { default as TextSectionBraceAside } from "./TextSection/TextSectionBraceAside";
+export { default as TextSectionQuoteAside } from "./TextSection/TextSectionQuoteAside";
 export { default as Button } from "./Buttons/Button";
 export { default as NavIconButton } from "./Buttons/NavIconButton";
 export { default as InputField } from "./forms/InputField";
