+++ src/pages/About/About.jsx
@@ -12,6 +12,7 @@ function About() {
         intro={tMission.intro}
         paragraphs={[tMission.body]}
         paragraphClassName="text-body text-grey-700 text-pretty"
+        contentRowClassName="md:items-center md:basis-3/5 xl:basis-1/2"
         aside={
           <TextSectionQuoteAside
             quote={tMission.quote}
@@ -19,6 +20,7 @@ function About() {
             initials={tMission.initials}
           />
         }
+        asideClassName="w-full shrink-0 md:basis-2/5 xl:basis-1/2 mt-6 md:max-w-md lg:max-w-lg xl:max-w-xl"
       />
       <div className="flex flex-col baseline-start gap-4">
         {/* <img src="/src/assets/placeholder-img.png" alt="placeholder image" /> */}
