+++ src/styles/index.css
@@ -75,7 +75,6 @@
   --color-primary-400: #4093f2;
   --color-primary-500: #2d66ff;
   --color-primary-600: #2247ad;
-  --color-primary-650: #193c9d;
   --color-primary-700: #1b3184;
   --color-primary-800: #043065;
   --color-primary-900: #032349;
